requirements.txt
openrgb-python==0.2.15
psutil==5.9.5
